+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - PIM-10619: Avoiding checks on expiration dates of certificates when disabling SSO
 - PIM-10621: Fix API returning an empty list instead of an empty object for published products
 - PIM-10613: Fix no purge done on the DQI when deleting option for multi-select and simple-select attributes
+- PIM-10627: Fix Job executions permissions for 'mass_edit','quick_export','mass_delete'
 
 ## Improvements